backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic[email]>=2.8,<3.0
PyJWT>=2.9,<3.0
passlib[bcrypt]>=1.7,<2.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
